Use standard cache dir as default cacheDirectory
Set the default cache directory as the
[common cache directory](https://github.com/avajs/find-cache-dir),
`./node_modules/.cache/babel-loader`.

Previously, when `cacheDirectory` was set to `true`, babel-loader
tried to use the operating system's temporary directory as a cache
directory.

However, when running npm scripts as the root user,
[npm overrides the TMPDIR environment variable](https://github.com/npm/npm/issues/4531).
This caused the cache files to be created in the project folder itself,
for example when using Docker: facebookincubator/create-react-app#483.