Add the message in BabelLoaderError's stack trace
When Error.captureStackTrace is called at the top of BabelLoaderError,
the error name and message are not added to the stack property. When
webpack wraps this in a ModuleBuildError, the message is lost.

Capturing at the end of the function ensures the message is part of the
stack trace, so it is eventually added to webpack's output.