fix(babel): Update to latest babel version and fix babelrc behaviour
This removes the custom behaviour in babel-loader of the option
`babelrc`. It now cannot be a string anymore and should be boolean and instead
extends should be used.

BREAKING CHANGE: Option `babelrc` needs to be boolean now. String values are not allowed anymore. Use `extends` option to specify a specific config file. A warning will be triggered when setting babelrc to a string.