Avoid absolute paths in cache filename hashes
By default, options will look something like:

```js
{
  inputSourceMap: undefined,
  sourceRoot: '/abs/path/to/project',
  filename: '/abs/path/to/project/src/index.js',
  sourceMap: false,
  sourceFileName: 'src/index.js',
}
```

The absolute paths here are problematic when running this on production
build machines that persist this cache but also have different absolute
paths. In this scenario, this causes the cache to be useless, and to
grow in size exceptionally fast.

Since the source filename has enough unique information here about
the cached file, and it is a relative version of the filename, we can
improve this by removing the absolute paths from the options before
hashing.